# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Know-me Doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/card": {
            "get": {
                "tags": [
                    "Card"
                ],
                "summary": "Index",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "qwe"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "qwer"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "qwer"
                                                    },
                                                    "alt": {
                                                        "type": "string",
                                                        "example": "qwe"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "qwe"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 4
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "admin"
                                                            },
                                                            "phone": {
                                                                "type": "string",
                                                                "example": null
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "skills": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "title": {
                                                                    "type": "string",
                                                                    "example": "skill"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://localhost:8878/api/v1/card?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://localhost:8878/api/v1/card?page=1"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "links": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "url": {
                                                                "type": "string",
                                                                "nullable": true
                                                            },
                                                            "label": {
                                                                "type": "string",
                                                                "example": "&laquo; Previous"
                                                            },
                                                            "active": {
                                                                "type": "boolean",
                                                                "example": false
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://localhost:8878/api/v1/card"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/card/{id}": {
            "get": {
                "tags": [
                    "Card"
                ],
                "summary": "Show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Card ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "qwe"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "qwer"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "qwer"
                                                },
                                                "alt": {
                                                    "type": "string",
                                                    "example": "qwe"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "qwe"
                                                },
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 4
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "admin"
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": null
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "skills": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "skill"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}